SearchBar: Minor refactor and moved to `blocks`.
